Sync AC_INIT package name with released tarballs
Released tarballs has the prefix "gst-plugins-at-acm", ie,
gst-plugins-at-acm_1.0.2.tar.gz.  But the package name defined in the
configure.ac doesn't have "-at" in it.

This does not do any bad.  It just that gst-inspect shows the pacakge
name as "Source module" in its output:

Plugin Details:
  Name                     acmh264dec
  Description              ACM H264 Decoder
  Filename                 /somewhere/src/.libs/libgstacmh264dec.so
  Version                  1.0.2
  License                  LGPL
  Source module            gst-plugins-acm

Just for the sake of completeness, sync them.